adapt to norminette

+++ playground/built-in/ft_export.c
@@ -1,10 +1,22 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_export.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ketomita <ketomita@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/10/07 10:07:14 by ketomita          #+#    #+#             */
+/*   Updated: 2025/10/07 10:14:38 by ketomita         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "builtin.h"
 
 static int	print_env_list(t_env *head)
 {
-	size_t	size;
+	int		size;
 	t_env	**env_array;
-	size_t	i;
+	int		i;
 
 	env_array = envcpy_and_get_size(head, &size);
 	if (!env_array)
