+++ playground/envlist/env_utils.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   env_utils.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ketomita <ketomita@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/10/07 10:18:00 by ketomita          #+#    #+#             */
+/*   Updated: 2025/10/07 10:18:01 by ketomita         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "envlist.h"
 
 void	free_env_list(t_env *head)
@@ -51,11 +63,11 @@ t_env	*find_env_node(t_env *head, const char *name)
 	return (NULL);
 }
 
-t_env	**envcpy_and_get_size(t_env *head, size_t *size)
+t_env	**envcpy_and_get_size(t_env *head, int *size)
 {
 	t_env	**env_array;
 	t_env	*current;
-	size_t	i;
+	int		i;
 
 	*size = 0;
 	current = head;
